+++ packages/sampler/stories/moonstone-stories/ContextualPopupDecorator.js
@@ -1,4 +1,5 @@
 import {ContextualPopupDecorator} from '@enact/moonstone/ContextualPopupDecorator';
+import BodyText from '@enact/moonstone/BodyText';
 import Button from '@enact/moonstone/Button';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
@@ -20,14 +21,16 @@ storiesOf('ContextualPopupDecorator')
 			<div style={{textAlign: 'center', marginTop: '100px'}}>
 				<ContextualButton
 					direction={select('direction', ['up', 'down', 'left', 'right'], 'down')}
-					onClick={action('onClick')}
-					onCloseButtonClick={action('onCloseButtonClick')}
+					noAutoDismiss={boolean('noAutoDismiss', false)}
+					onClose={action('onClose')}
 					open={boolean('open', false)}
-					showCloseButton={boolean('showCloseButton', false)}
 					popupComponent={renderPopup}
+					showCloseButton={boolean('showCloseButton', false)}
+					spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], 'self-first')}
 				>
 					{text('button string', 'Hello Contextual Button')}
 				</ContextualButton>
+				<BodyText centered>Use KNOBS to interact with the ContextualPopup.</BodyText>
 			</div>
 		)
 	);