+++ packages/moonstone/CHANGELOG.md
@@ -2,10 +2,26 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
-## [unreleased]
+## [1.0.0-beta.1] - 2016-12-30
+
+### Added
+
+- `Moonstone/VideoPlayer` and `moonstone/TooltipDecorator` components and samples
+- `moonstone/Panels.Panels` property `onBack` to support `ui/Cancelable`
+- `moonstone/VirtualFlexList` Work-In-Progress component to support variably sized rows or columns
+- `moonstone/ExpandableItem` properties `autoClose` and `lockBottom`
+- `moonstone/ExpandableList` properties `noAutoClose` and `noLockBottom`
+- `moonstone/ContextualPopup` property `noAutoDismiss`
+- `moonstone/Dialog` property `scrimType`
+- `moonstone/Popup` property `spotlightRestrict`
 
 ### Changed
 
+- `moonstone/Panels.Routable` to require a `navigate` configuration property indicating the event callback for back or cancel actions
+- `moonstone/MarqueeController` focus/blur handling to start and stop synchronized `moonstone/Marquee` components
+- `moonstone/ExpandableList` property `autoClose` to `closeOnSelect` to disambiguate it from the added `autoClose` on 5-way up
+- `moonstone/ContextualPopupDecorator.ContextualPopupDecorator` component's `onCloseButtonClick` property to `onClose`
+- `moonstone/Dialog` component's `onCloseButtonClicked` property to `onClose`
 - `moonstone/Spinner` component's `center` and `middle` properties to a single `centered` property
 	that applies both horizontal and vertical centering
 - `moonstone/Popup.PopupBase` component's `onCloseButtonClicked` property to `onCloseButtonClick`
@@ -13,11 +29,28 @@ The following is a curated list of changes in the Enact moonstone module, newest
 	effect can be achieved by omitting the property or passing `null`.
 - `moonstone/VirtualGridList` to be scrolled by page when navigating with a 5-way direction key
 - `moonstone/Scroller`, `moonstone/VirtualList`, `moonstone/VirtualGridList`, and `moonstone/Scrollable` to no longer respond to mouse down/move/up events
+- all Expandables to include a state arrow UI element
+- `moonstone/LabeledItem` to support a `titleIcon` property which positions just after the title text
+- `moonstone/Button` to include `moonstone/TooltipDecorator`
+- `moonstone/Expandable` to support being managed, radio group-style, by a component wrapped with `RadioControllerDecorator` from `ui/RadioDecorator`
+- `moonstone/Picker` to animate `moonstone/Marquee` children when any part of the `moonstone/Picker` is focused
+- `moonstone/VirtualList` to mute its container instead of disabling it during scroll events
+- `moonstone/VirtualList`, `moonstone/VirtualGridList`, and `moonstone/Scroller` to continue scrolling when holding down the paging controls
+- `moonstone/VirtualList` to require a `component` prop and not have a default value
+- `moonstone/Picker` to continuously change when a button is held down by adding `ui/Holdable`.
 
 ### Fixed
 
+- `moonstone/Popup` and `moonstone/ContextualPopup` 5-way navigation behavior using spotlight.
+- Bug where a synchronized marquee whose content fit the available space would prevent restarting of the marquees
 - `moonstone/Input` to show an ellipsis on the correct side based on the text directionality of the `value` or `placeholder` content.
 - `moonstone/VirtualList` and `moonstone/VirtualGridList` to prevent unwanted scrolling when focused with the pointer
+- `moonstone/Picker` to remove fingernail when a the pointer is held down, but the pointer is moved off the `joined` picker.
+- `moonstone/LabeledItem` to include marquee on both `title` and `label`, and be synchronized
+
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
 
 ## [1.0.0-alpha.4] - 2016-12-2