+++ packages/ui/CHANGELOG.md
@@ -4,17 +4,31 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ## [unreleased]
 
-### Added
-
-- `none` to `scrimType` prop values in `ui/FloatingLayer`
-
 ### Changed
 
 - `ui/Pressable` to also accept an array of events for depress/release, to support multiple interaction methods
 
+## [1.0.0-beta.1] - 2016-12-30
+
+### Added
+
+- `ui/RadioDecorator` and `ui/RadioControllerDecorator` to support radio group-style management of components
+- `ui/Holdable` Higher-order Component
+- `ui/ViewManager` events `onAppear`, `onEnter`, `onLeave`, `onStay`, `onTransition`, and `onWillTransition`
+- `ui/FloatingLayer` `scrimType` prop value `none`
+- `ui/Pressable` config option `onMouseLeave`
+
 ### Removed
 
-- `fit` prop from `ui/Transition` in favor of using `className`
+- `ui/Transition` prop `fit` in favor of using `className`
+
+### Changed
+
+- `ui/FloatingLayer` property `autoDismiss` to handle both ESC key and click events
+
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
 
 ## [1.0.0-alpha.4] - 2016-12-2